5d210a2 [REFACTOR] ./model/* : Refactoring

+++ src/main/java/com/pos/app/model/DaySalesData.java
@@ -11,6 +11,6 @@ public class DaySalesData {
     private Integer id;
     private Integer totalInvoicedOrders;
     private Integer totalItemsSold;
-    private Integer totalRevenueGenerated;
+    private Double totalRevenueGenerated;
     private ZonedDateTime reportDate;
 }
